% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{create_api_key}
\alias{create_api_key}
\title{Creates a unique key that you can distribute to clients who are executing your API}
\usage{
create_api_key(apiId, description = NULL, expires = NULL)
}
\arguments{
\item{apiId}{[required] The ID for your GraphQL API.}

\item{description}{A description of the purpose of the API key.}

\item{expires}{The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .}
}
\description{
Creates a unique key that you can distribute to clients who are executing your API.
}
\section{Accepted Parameters}{
\preformatted{create_api_key(
  apiId = "string",
  description = "string",
  expires = 123
)
}
}

