% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{create_data_source}
\alias{create_data_source}
\title{Creates a \code{DataSource} object}
\usage{
create_data_source(apiId, name, type, description = NULL,
  serviceRoleArn = NULL, dynamodbConfig = NULL, lambdaConfig = NULL,
  elasticsearchConfig = NULL, httpConfig = NULL)
}
\arguments{
\item{apiId}{The API ID for the GraphQL API for the \code{DataSource}.}

\item{name}{A user-supplied name for the \code{DataSource}.}

\item{type}{The type of the \code{DataSource}.}

\item{description}{A description of the \code{DataSource}.}

\item{serviceRoleArn}{The IAM service role ARN for the data source. The system assumes this role when accessing the data source.}

\item{dynamodbConfig}{DynamoDB settings.}

\item{lambdaConfig}{AWS Lambda settings.}

\item{elasticsearchConfig}{Amazon Elasticsearch settings.}

\item{httpConfig}{Http endpoint settings.}
}
\description{
Creates a \code{DataSource} object.
}
\examples{

}
