% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{create_resolver}
\alias{create_resolver}
\title{Creates a Resolver object}
\usage{
create_resolver(apiId, typeName, fieldName, dataSourceName,
  requestMappingTemplate, responseMappingTemplate = NULL)
}
\arguments{
\item{apiId}{[required] The ID for the GraphQL API for which the resolver is being created.}

\item{typeName}{[required] The name of the \code{Type}.}

\item{fieldName}{[required] The name of the field to attach the resolver to.}

\item{dataSourceName}{[required] The name of the data source for which the resolver is being created.}

\item{requestMappingTemplate}{[required] The mapping template to be used for requests.

A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL).}

\item{responseMappingTemplate}{The mapping template to be used for responses from the data source.}
}
\description{
Creates a \code{Resolver} object.
}
\details{
A resolver converts incoming requests into a format that a data source can understand and converts the data source's responses into GraphQL.
}
\section{Accepted Parameters}{
\preformatted{create_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string",
  dataSourceName = "string",
  requestMappingTemplate = "string",
  responseMappingTemplate = "string"
)
}
}

