% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{create_type}
\alias{create_type}
\title{Creates a \code{Type} object}
\usage{
create_type(
  apiId = "string",
  definition = "string",
  format = "SDL"|"JSON"
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{definition}{[required] The type definition, in GraphQL Schema Definition Language (SDL) format.

For more information, see the \href{http://graphql.org/learn/schema/}{GraphQL SDL documentation}.}

\item{format}{[required] The type format: SDL or JSON.}
}
\description{
Creates a \code{Type} object.
}
\examples{

}
