% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_function}
\alias{delete_function}
\title{Deletes a \code{Function}}
\usage{
delete_function(
  apiId = "string",
  functionId = "string"
)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{functionId}{[required] The \code{Function} ID.}
}
\description{
Deletes a \code{Function}.
}
\examples{

}
