% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_graphql_api}
\alias{delete_graphql_api}
\title{Deletes a GraphqlApi object}
\usage{
delete_graphql_api(apiId)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\description{
Deletes a \code{GraphqlApi} object.
}
\section{Accepted Parameters}{
\preformatted{delete_graphql_api(
  apiId = "string"
)
}
}

