% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_type}
\alias{delete_type}
\title{Deletes a Type object}
\usage{
delete_type(apiId, typeName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The type name.}
}
\description{
Deletes a \code{Type} object.
}
\section{Accepted Parameters}{
\preformatted{delete_type(
  apiId = "string",
  typeName = "string"
)
}
}

