% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_function}
\alias{get_function}
\title{Get a Function}
\usage{
get_function(apiId, functionId)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{functionId}{[required] The \code{Function} ID.}
}
\description{
Get a \code{Function}.
}
\section{Accepted Parameters}{
\preformatted{get_function(
  apiId = "string",
  functionId = "string"
)
}
}

