% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_graphql_api}
\alias{get_graphql_api}
\title{Retrieves a \code{GraphqlApi} object}
\usage{
get_graphql_api(apiId)
}
\arguments{
\item{apiId}{The API ID for the GraphQL API.}
}
\description{
Retrieves a \code{GraphqlApi} object.
}
\examples{

}
