% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_introspection_schema}
\alias{get_introspection_schema}
\title{Retrieves the introspection schema for a GraphQL API}
\usage{
get_introspection_schema(
  apiId = "string",
  format = "SDL"|"JSON"
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{format}{[required] The schema format: SDL or JSON.}
}
\description{
Retrieves the introspection schema for a GraphQL API.
}
\examples{

}
