% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_resolver}
\alias{get_resolver}
\title{Retrieves a \code{Resolver} object}
\usage{
get_resolver(apiId, typeName, fieldName)
}
\arguments{
\item{apiId}{The API ID.}

\item{typeName}{The resolver type name.}

\item{fieldName}{The resolver field name.}
}
\description{
Retrieves a \code{Resolver} object.
}
\examples{

}
