% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_schema_creation_status}
\alias{get_schema_creation_status}
\title{Retrieves the current status of a schema creation operation}
\usage{
get_schema_creation_status(
  apiId = "string"
)
}
\arguments{
\item{apiId}{[required] The API ID.}
}
\description{
Retrieves the current status of a schema creation operation.
}
\examples{

}
