% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_type}
\alias{get_type}
\title{Retrieves a \code{Type} object}
\usage{
get_type(apiId, typeName, format)
}
\arguments{
\item{apiId}{The API ID.}

\item{typeName}{The type name.}

\item{format}{The type format: SDL or JSON.}
}
\description{
Retrieves a \code{Type} object.
}
\examples{

}
