% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{list_api_keys}
\alias{list_api_keys}
\title{Lists the API keys for a given API}
\usage{
list_api_keys(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
Lists the API keys for a given API.
}
\details{
API keys are deleted automatically sometime after they expire. However, they may still be included in the response until they have actually been deleted. You can safely call \code{DeleteApiKey} to manually delete a key before it's automatically deleted.
}
\examples{

}
