% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{list_data_sources}
\alias{list_data_sources}
\title{Lists the data sources for a given API}
\usage{
list_data_sources(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
Lists the data sources for a given API.
}
\examples{

}
