% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{list_functions}
\alias{list_functions}
\title{List multiple functions}
\usage{
list_functions(apiId, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
List multiple functions.
}
\section{Accepted Parameters}{
\preformatted{list_functions(
  apiId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

