% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{list_graphql_apis}
\alias{list_graphql_apis}
\title{Lists your GraphQL APIs}
\usage{
list_graphql_apis(
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{nextToken}{An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
Lists your GraphQL APIs.
}
\examples{

}
