% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{list_resolvers_by_function}
\alias{list_resolvers_by_function}
\title{List the resolvers that are associated with a specific function}
\usage{
list_resolvers_by_function(
  apiId = "string",
  functionId = "string",
  nextToken = "string",
  maxResults = 123
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{functionId}{[required] The Function ID.}

\item{nextToken}{An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.}

\item{maxResults}{The maximum number of results you want the request to return.}
}
\description{
List the resolvers that are associated with a specific function.
}
\examples{

}
