% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_package.R
\docType{package}
\name{paws.appsync-package}
\alias{paws.appsync}
\alias{paws.appsync-package}
\title{paws.appsync: AWS AppSync}
\description{
AWS AppSync provides API actions for creating and
interacting with data sources using GraphQL from your application.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
