% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{start_schema_creation}
\alias{start_schema_creation}
\title{Adds a new schema to your GraphQL API}
\usage{
start_schema_creation(apiId, definition)
}
\arguments{
\item{apiId}{The API ID.}

\item{definition}{The schema definition, in GraphQL schema language format.}
}
\description{
Adds a new schema to your GraphQL API.
}
\details{
This operation is asynchronous. Use to determine when it has completed.
}
\examples{

}
