% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_api_key}
\alias{update_api_key}
\title{Updates an API key}
\usage{
update_api_key(
  apiId = "string",
  id = "string",
  description = "string",
  expires = 123
)
}
\arguments{
\item{apiId}{[required] The ID for the GraphQL API.}

\item{id}{[required] The API key ID.}

\item{description}{A description of the purpose of the API key.}

\item{expires}{The time from update time after which the API key expires. The date is represented as seconds since the epoch. For more information, see .}
}
\description{
Updates an API key.
}
\examples{

}
