% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_function}
\alias{update_function}
\title{Updates a \code{Function} object}
\usage{
update_function(
  apiId = "string",
  name = "string",
  description = "string",
  functionId = "string",
  dataSourceName = "string",
  requestMappingTemplate = "string",
  responseMappingTemplate = "string",
  functionVersion = "string"
)
}
\arguments{
\item{apiId}{[required] The GraphQL API ID.}

\item{name}{[required] The \code{Function} name.}

\item{description}{The \code{Function} description.}

\item{functionId}{[required] The function ID.}

\item{dataSourceName}{[required] The \code{Function} \code{DataSource} name.}

\item{requestMappingTemplate}{[required] The \code{Function} request mapping template. Functions support only the 2018-05-29 version of the request mapping template.}

\item{responseMappingTemplate}{The \code{Function} request mapping template.}

\item{functionVersion}{[required] The \code{version} of the request mapping template. Currently the supported value is 2018-05-29.}
}
\description{
Updates a \code{Function} object.
}
\examples{

}
