% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_resolver}
\alias{update_resolver}
\title{Updates a Resolver object}
\usage{
update_resolver(apiId, typeName, fieldName, dataSourceName,
  requestMappingTemplate, responseMappingTemplate = NULL)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The new type name.}

\item{fieldName}{[required] The new field name.}

\item{dataSourceName}{[required] The new data source name.}

\item{requestMappingTemplate}{[required] The new request mapping template.}

\item{responseMappingTemplate}{The new response mapping template.}
}
\description{
Updates a \code{Resolver} object.
}
\section{Accepted Parameters}{
\preformatted{update_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string",
  dataSourceName = "string",
  requestMappingTemplate = "string",
  responseMappingTemplate = "string"
)
}
}

