% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_type}
\alias{update_type}
\title{Updates a \code{Type} object}
\usage{
update_type(
  apiId = "string",
  typeName = "string",
  definition = "string",
  format = "SDL"|"JSON"
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The new type name.}

\item{definition}{The new definition.}

\item{format}{[required] The new type format: SDL or JSON.}
}
\description{
Updates a \code{Type} object.
}
\examples{

}
