% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{create_graphql_api}
\alias{create_graphql_api}
\title{Creates a \code{GraphqlApi} object}
\usage{
create_graphql_api(name, authenticationType, logConfig = NULL,
  userPoolConfig = NULL, openIDConnectConfig = NULL)
}
\arguments{
\item{name}{A user-supplied name for the \code{GraphqlApi}.}

\item{authenticationType}{The authentication type: API key, IAM, or Amazon Cognito User Pools.}

\item{logConfig}{The Amazon CloudWatch logs configuration.}

\item{userPoolConfig}{The Amazon Cognito User Pool configuration.}

\item{openIDConnectConfig}{The Open Id Connect configuration configuration.}
}
\description{
Creates a \code{GraphqlApi} object.
}
\examples{

}
