% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_api_key}
\alias{delete_api_key}
\title{Deletes an API key}
\usage{
delete_api_key(apiId, id)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{id}{[required] The ID for the API key.}
}
\description{
Deletes an API key.
}
\section{Accepted Parameters}{
\preformatted{delete_api_key(
  apiId = "string",
  id = "string"
)
}
}

