% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_data_source}
\alias{delete_data_source}
\title{Deletes a \code{DataSource} object}
\usage{
delete_data_source(apiId, name)
}
\arguments{
\item{apiId}{The API ID.}

\item{name}{The name of the data source.}
}
\description{
Deletes a \code{DataSource} object.
}
\examples{

}
