% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_graphql_api}
\alias{delete_graphql_api}
\title{Deletes a \code{GraphqlApi} object}
\usage{
delete_graphql_api(apiId)
}
\arguments{
\item{apiId}{The API ID.}
}
\description{
Deletes a \code{GraphqlApi} object.
}
\examples{

}
