% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_resolver}
\alias{delete_resolver}
\title{Deletes a Resolver object}
\usage{
delete_resolver(apiId, typeName, fieldName)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The name of the resolver type.}

\item{fieldName}{[required] The resolver field name.}
}
\description{
Deletes a \code{Resolver} object.
}
\section{Accepted Parameters}{
\preformatted{delete_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string"
)
}
}

