% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{delete_type}
\alias{delete_type}
\title{Deletes a \code{Type} object}
\usage{
delete_type(apiId, typeName)
}
\arguments{
\item{apiId}{The API ID.}

\item{typeName}{The type name.}
}
\description{
Deletes a \code{Type} object.
}
\examples{

}
