% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_data_source}
\alias{get_data_source}
\title{Retrieves a DataSource object}
\usage{
get_data_source(apiId, name)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{name}{[required] The name of the data source.}
}
\description{
Retrieves a \code{DataSource} object.
}
\section{Accepted Parameters}{
\preformatted{get_data_source(
  apiId = "string",
  name = "string"
)
}
}

