% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_graphql_api}
\alias{get_graphql_api}
\title{Retrieves a GraphqlApi object}
\usage{
get_graphql_api(apiId)
}
\arguments{
\item{apiId}{[required] The API ID for the GraphQL API.}
}
\description{
Retrieves a \code{GraphqlApi} object.
}
\section{Accepted Parameters}{
\preformatted{get_graphql_api(
  apiId = "string"
)
}
}

