% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_resolver}
\alias{get_resolver}
\title{Retrieves a \code{Resolver} object}
\usage{
get_resolver(
  apiId = "string",
  typeName = "string",
  fieldName = "string"
)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The resolver type name.}

\item{fieldName}{[required] The resolver field name.}
}
\description{
Retrieves a \code{Resolver} object.
}
\examples{

}
