% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{get_type}
\alias{get_type}
\title{Retrieves a Type object}
\usage{
get_type(apiId, typeName, format)
}
\arguments{
\item{apiId}{[required] The API ID.}

\item{typeName}{[required] The type name.}

\item{format}{[required] The type format: SDL or JSON.}
}
\description{
Retrieves a \code{Type} object.
}
\section{Accepted Parameters}{
\preformatted{get_type(
  apiId = "string",
  typeName = "string",
  format = "SDL"|"JSON"
)
}
}

