% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_data_source}
\alias{update_data_source}
\title{Updates a \code{DataSource} object}
\usage{
update_data_source(apiId, name, type, description = NULL,
  serviceRoleArn = NULL, dynamodbConfig = NULL, lambdaConfig = NULL,
  elasticsearchConfig = NULL, httpConfig = NULL)
}
\arguments{
\item{apiId}{The API ID.}

\item{name}{The new name for the data source.}

\item{type}{The new data source type.}

\item{description}{The new description for the data source.}

\item{serviceRoleArn}{The new service role ARN for the data source.}

\item{dynamodbConfig}{The new DynamoDB configuration.}

\item{lambdaConfig}{The new Lambda configuration.}

\item{elasticsearchConfig}{The new Elasticsearch configuration.}

\item{httpConfig}{The new http endpoint configuration}
}
\description{
Updates a \code{DataSource} object.
}
\examples{

}
