% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_resolver}
\alias{update_resolver}
\title{Updates a \code{Resolver} object}
\usage{
update_resolver(apiId, typeName, fieldName, dataSourceName,
  requestMappingTemplate, responseMappingTemplate = NULL)
}
\arguments{
\item{apiId}{The API ID.}

\item{typeName}{The new type name.}

\item{fieldName}{The new field name.}

\item{dataSourceName}{The new data source name.}

\item{requestMappingTemplate}{The new request mapping template.}

\item{responseMappingTemplate}{The new response mapping template.}
}
\description{
Updates a \code{Resolver} object.
}
\examples{

}
