% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.appsync_operations.R
\name{update_type}
\alias{update_type}
\title{Updates a \code{Type} object}
\usage{
update_type(apiId, typeName, format, definition = NULL)
}
\arguments{
\item{apiId}{The API ID.}

\item{typeName}{The new type name.}

\item{format}{The new type format: SDL or JSON.}

\item{definition}{The new definition.}
}
\description{
Updates a \code{Type} object.
}
\examples{

}
